requirements.txt
chatterbox-tts
fastapi
uvicorn
pydantic
torchaudio
torch
torchvision
torchaudio